Added uart (wireless) interface. The code related to the different interfaces now reside in different files and the old external_interface.c just call the appropriate functions depending on board settings. Currently, the uart interface work identically and in parallel to the usb interface. It has it's own communication buffers and logics. The communication that is requested over each interface is delivered back on the same interface. However, the processing is common so both interfaces will manipulate the same processing sequence.
The uart communication is handled with interrupts. In comparison to the USB is has limited bandwidth. In case the user requrests more than it can handle, it will start dropping bytes.

The intention is to rewrite the uart (wireless) communication logic such that it suits a lossy wireless link better. In the end it will not be intended for logging all raw data but rather for outputting processed data, e.g. step displacement.

Signed-off-by: John-Olof Nilsson <jnil02@kth.se>